Update FirstPost.md
Reviewer 2:
What brands were new to Estes Park?
What was your vision for the business? Both visually in the store, and for the business as a whole.
Overall: 
What will the 3 other posts be? 
Is your audience the people in Estes Park? So they can learn the knowledge and vision of the new Hyk business in the town? 
The purpose of this blog is...?